+++ src/endpoint/routes.ts
@@ -3,10 +3,8 @@ import Joi from 'joi'
 import boom from '@hapi/boom'
 import { logger } from '@libp2p/logger'
 import { tmpDir } from '../utils.js'
-
-/**
- * @typedef {import('../types').Factory} Factory
- */
+import type { Server } from '@hapi/hapi'
+import type { Factory } from '../index.js'
 
 const debug = logger('ipfsd-ctl:routes')
 
@@ -18,12 +16,9 @@ const routeOptions = {
   }
 }
 
-/**
- * @param {Error & { stdout?: string }} err
- */
-const badRequest = err => {
+const badRequest = (err: Error & { stdout?: string }) => {
   let msg
-  if (err.stdout) {
+  if (err.stdout != null) {
     msg = err.stdout + ' - ' + err.message
   } else {
     msg = err.message
@@ -32,27 +27,17 @@ const badRequest = err => {
   throw boom.badRequest(msg)
 }
 
-/**
- * @type {Record<string, any>}
- */
-const nodes = {}
+const nodes: Record<string, any> = {}
 
-/**
- * @namespace EndpointServerRoutes
- * @ignore
- * @param {import('@hapi/hapi').Server} server
- * @param {() => Factory | Promise<Factory>} createFactory
- * @returns {void}
- */
-export default (server, createFactory) => {
+export default (server: Server, createFactory: () => Factory | Promise<Factory>): void => {
   server.route({
     method: 'GET',
     path: '/util/tmp-dir',
     handler: async (request) => {
-      const type = request.query.type || 'go'
+      const type = request.query.type ?? 'go'
       try {
         return { tmpDir: await tmpDir(type) }
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     }
@@ -66,7 +51,7 @@ export default (server, createFactory) => {
 
       try {
         return { version: await nodes[id].version() }
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     },
@@ -77,7 +62,7 @@ export default (server, createFactory) => {
     method: 'POST',
     path: '/spawn',
     handler: async (request) => {
-      const opts = request.payload || {}
+      const opts = request.payload ?? {}
       try {
         const ipfsd = await createFactory()
         const id = nanoid()
@@ -85,9 +70,9 @@ export default (server, createFactory) => {
         nodes[id] = await ipfsd.spawn(opts)
         return {
           id: id,
-          apiAddr: nodes[id].apiAddr ? nodes[id].apiAddr.toString() : '',
-          gatewayAddr: nodes[id].gatewayAddr ? nodes[id].gatewayAddr.toString() : '',
-          grpcAddr: nodes[id].grpcAddr ? nodes[id].grpcAddr.toString() : '',
+          apiAddr: nodes[id].apiAddr?.toString(),
+          gatewayAddr: nodes[id].gatewayAddr?.toString(),
+          grpcAddr: nodes[id].grpcAddr?.toString(),
           initialized: nodes[id].initialized,
           started: nodes[id].started,
           disposable: nodes[id].disposable,
@@ -95,7 +80,7 @@ export default (server, createFactory) => {
           path: nodes[id].path,
           clean: nodes[id].clean
         }
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     }
@@ -109,7 +94,7 @@ export default (server, createFactory) => {
     path: '/init',
     handler: async (request) => {
       const id = request.query.id
-      const payload = request.payload || {}
+      const payload = request.payload ?? {}
 
       try {
         await nodes[id].init(payload)
@@ -117,7 +102,7 @@ export default (server, createFactory) => {
         return {
           initialized: nodes[id].initialized
         }
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     },
@@ -137,11 +122,11 @@ export default (server, createFactory) => {
         await nodes[id].start()
 
         return {
-          apiAddr: nodes[id].apiAddr ? nodes[id].apiAddr.toString() : '',
-          gatewayAddr: nodes[id].gatewayAddr ? nodes[id].gatewayAddr.toString() : '',
-          grpcAddr: nodes[id].grpcAddr ? nodes[id].grpcAddr.toString() : ''
+          apiAddr: nodes[id].apiAddr?.toString(),
+          gatewayAddr: nodes[id].gatewayAddr?.toString(),
+          grpcAddr: nodes[id].grpcAddr?.toString()
         }
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     },
@@ -163,7 +148,7 @@ export default (server, createFactory) => {
         await nodes[id].cleanup()
 
         return h.response().code(200)
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     },
@@ -183,7 +168,7 @@ export default (server, createFactory) => {
         await nodes[id].stop()
 
         return h.response().code(200)
-      } catch (/** @type {any} */ err) {
+      } catch (err: any) {
         badRequest(err)
       }
     },