+++ src/index.ts
@@ -1,32 +1,55 @@
-
-import type { EventEmitter } from 'events'
+import DefaultFactory from './factory.js'
+import Server from './endpoint/server.js'
 import type { IPFS } from 'ipfs-core-types'
 import type { Multiaddr } from '@multiformats/multiaddr'
 import type { PeerId } from '@libp2p/interface-peer-id'
-
-export interface Subprocess {
-  stderr: EventEmitter | null
-  stdout: EventEmitter | null
-}
+import type { ExecaChildProcess } from 'execa'
 
 export interface PeerData {
   id: PeerId
   addresses: Multiaddr[]
 }
 
-export interface Controller {
-  init: (options?: InitOptions) => Promise<Controller>
-  start: () => Promise<Controller>
-  stop: () => Promise<Controller>
-  cleanup: () => Promise<Controller>
+export type ControllerType = 'js' | 'go' | 'proc'
+
+export interface Controller<Type extends ControllerType = 'go'> {
+  /**
+   * Initialize a repo
+   */
+  init: (options?: InitOptions) => Promise<Controller<Type>>
+
+  /**
+   * Start the daemon
+   */
+  start: () => Promise<Controller<Type>>
+
+  /**
+   * Stop the daemon
+   */
+  stop: () => Promise<Controller<Type>>
+
+  /**
+   * Delete the repo that was being used.
+   * If the node was marked as `disposable` this will be called
+   * automatically when the process is exited.
+   */
+  cleanup: () => Promise<Controller<Type>>
+
+  /**
+   * Get the pid of the `ipfs daemon` process
+   */
   pid: () => Promise<number>
+
+  /**
+   * Get the version of ipfs
+   */
   version: () => Promise<string>
   path: string
   started: boolean
   initialized: boolean
   clean: boolean
-  api: IPFS
-  subprocess?: Subprocess | null
+  api: IPFSAPI
+  subprocess?: ExecaChildProcess | null
   opts: ControllerOptions
   apiAddr: Multiaddr
   peer: PeerData
@@ -43,8 +66,6 @@ export interface RemoteState {
   gatewayAddr: string
   grpcAddr: string
 }
-
-export type NodeType = 'js' | 'go' | 'proc'
 
 export interface InitOptions {
   pass?: string
@@ -135,7 +156,7 @@ export interface IPFSOptions {
   repoAutoMigrate?: boolean
 }
 
-export interface ControllerOptions {
+export interface ControllerOptions<Type extends ControllerType = ControllerType> {
   /**
    * Flag to activate custom config for tests
    */
@@ -155,7 +176,7 @@ export interface ControllerOptions {
   /**
    * The daemon type
    */
-  type?: NodeType
+  type?: Type
   /**
    * Additional environment variables, passed to executing shell. Only applies for Daemon controllers
    */
@@ -168,6 +189,10 @@ export interface ControllerOptions {
    * Reference to an ipfs-http-client module
    */
   ipfsHttpModule?: any
+  /**
+   * Reference to a kubo-rpc-client module
+   */
+  kuboRpcModule?: any
   /**
    * Reference to an ipfs or ipfs-core module
    */
@@ -195,15 +220,69 @@ export interface ControllerOptions {
 }
 
 export interface ControllerOptionsOverrides {
-  js?: ControllerOptions
-  go?: ControllerOptions
-  proc?: ControllerOptions
+  js?: ControllerOptions<'js'>
+  go?: ControllerOptions<'go'>
+  proc?: ControllerOptions<'proc'>
 }
 
-export interface Factory {
+export interface Factory<Type extends ControllerType = ControllerType> {
   tmpDir: (options?: ControllerOptions) => Promise<string>
-  spawn: (options?: ControllerOptions) => Promise<Controller>
+  spawn: (options?: ControllerOptions) => Promise<Controller<Type>>
   clean: () => Promise<void>
-  controllers: Controller[]
-  opts: ControllerOptions
+  controllers: Array<Controller<Type>>
+  opts: ControllerOptions<Type>
+}
+
+export interface CreateFactory { (): Factory | Promise<Factory> }
+
+/**
+ * Creates a factory
+ *
+ * @param {ControllerOptions} [options]
+ * @param {ControllerOptionsOverrides} [overrides]
+ * @returns {Factory}
+ */
+export const createFactory = (options?: ControllerOptions, overrides?: ControllerOptionsOverrides): Factory => {
+  return new DefaultFactory(options, overrides)
+}
+
+/**
+ * Creates a node
+ */
+export const createController = async (options?: ControllerOptions): Promise<Controller> => {
+  const f = new DefaultFactory()
+  return await f.spawn(options)
+}
+
+export interface IPFSAPI extends IPFS {
+  apiHost?: string
+  apiPort?: number
+  gatewayHost?: string
+  gatewayPort?: number
+  grpcHost?: string
+  grpcPort?: number
+}
+
+/**
+ * Create a Endpoint Server
+ *
+ * @param {number | { port: number }} [options] - Configuration options or just the port.
+ * @param {ControllerOptions} [factoryOptions]
+ * @param {ControllerOptionsOverrides} [factoryOverrides]
+ */
+export const createServer = (options?: number | { port: number }, factoryOptions: ControllerOptions = {}, factoryOverrides: ControllerOptionsOverrides = {}) => {
+  let port: number | undefined
+
+  if (typeof options === 'number') {
+    port = options
+  } else if (options != null) {
+    port = options.port
+  }
+
+  return new Server({
+    port,
+    host: '127.0.0.1'
+  }, () => {
+    return createFactory(factoryOptions, factoryOverrides)
+  })
 }