+++ test/create.spec.ts
@@ -1,22 +1,20 @@
 /* eslint-env mocha */
+/* eslint-disable @typescript-eslint/restrict-template-expressions */
 
 import { expect } from 'aegir/chai'
 import { isNode, isBrowser, isWebWorker } from 'wherearewe'
-import { createFactory, createController, createServer } from '../src/index.js'
+import { createFactory, createController, createServer, ControllerOptions } from '../src/index.js'
 import Client from '../src/ipfsd-client.js'
 import Daemon from '../src/ipfsd-daemon.js'
 import Proc from '../src/ipfsd-in-proc.js'
 import * as ipfsModule from 'ipfs'
 import * as ipfsHttpModule from 'ipfs-http-client'
-// @ts-ignore no types
+// @ts-expect-error no types
 import * as goIpfsModule from 'go-ipfs'
 import * as ipfsClientModule from 'ipfs-client'
+import * as kuboRpcModule from 'kubo-rpc-client'
 
-/**
- * @typedef {import('../src/types').ControllerOptions} ControllerOptions
- */
-
-describe('`createController` should return the correct class', async () => {
+describe('`createController` should return the correct class', () => {
   it('for type `js` ', async () => {
     const f = await createController({
       type: 'js',
@@ -36,7 +34,7 @@ describe('`createController` should return the correct class', async () => {
     const f = await createController({
       type: 'go',
       disposable: false,
-      ipfsHttpModule,
+      kuboRpcModule,
       ipfsBin: isNode ? goIpfsModule.path() : undefined
     })
 
@@ -73,20 +71,14 @@ describe('`createController` should return the correct class', async () => {
       disposable: false,
       ipfsModule,
       ipfsClientModule: {
-        /**
-         * @param {any} opts
-         */
-        create: (opts) => {
+        create: (opts: any) => {
           clientCreated = true
 
           return ipfsClientModule.create(opts)
         }
       },
       ipfsHttpModule: {
-        /**
-         * @param {any} opts
-         */
-        create: async (opts) => {
+        create: async (opts: any) => {
           httpCreated = true
 
           return ipfsHttpModule.create(opts)
@@ -108,20 +100,14 @@ describe('`createController` should return the correct class', async () => {
       disposable: false,
       ipfsModule,
       ipfsClientModule: {
-        /**
-         * @param {any} opts
-         */
-        create: (opts) => {
+        create: (opts: any) => {
           clientCreated = true
 
           return ipfsClientModule.create(opts)
         }
       },
       ipfsHttpModule: {
-        /**
-         * @param {any} opts
-         */
-        create: async (opts) => {
+        create: async (opts: any) => {
           httpCreated = true
 
           return ipfsHttpModule.create(opts)
@@ -136,38 +122,33 @@ describe('`createController` should return the correct class', async () => {
   })
 })
 
-const defaultOps = {
-  ipfsHttpModule
-}
-
-/** @type {ControllerOptions[]} */
-const types = [{
-  ...defaultOps,
+const types: ControllerOptions[] = [{
   type: 'js',
+  ipfsHttpModule,
   test: true,
   ipfsModule,
   ipfsBin: isNode ? ipfsModule.path() : undefined
 }, {
-  ...defaultOps,
   ipfsBin: isNode ? goIpfsModule.path() : undefined,
   type: 'go',
+  kuboRpcModule,
   test: true
 }, {
-  ...defaultOps,
   type: 'proc',
+  ipfsHttpModule,
   test: true,
   ipfsModule
 }, {
-  ...defaultOps,
   type: 'js',
+  ipfsHttpModule,
   test: true,
   remote: true,
   ipfsModule,
   ipfsBin: isNode ? ipfsModule.path() : undefined
 }, {
-  ...defaultOps,
   ipfsBin: isNode ? goIpfsModule.path() : undefined,
   type: 'go',
+  kuboRpcModule,
   test: true,
   remote: true
 }]